Move inputObserver
