+++ packages/rum/src/domain/record/record.ts
@@ -16,13 +16,12 @@ import type {
 import { RecordType, IncrementalSource } from '../../types'
 import { serializeDocument, SerializationContextStatus } from './serialize'
 import { initObservers } from './observers/observers'
-import type { InputCallback } from './observers/observers'
-
 import { getVisualViewport, getScrollX, getScrollY } from './viewports'
 import { assembleIncrementalSnapshot } from './utils'
 import { createElementsScrollPositions } from './elementsScrollPositions'
 import type { ShadowRootsController } from './shadowRootsController'
 import { initShadowRootsController } from './shadowRootsController'
+import type { InputCallback } from './observers/inputObserver'
 
 export interface RecordOptions {
   emit?: (record: BrowserRecord) => void
