+++ packages/rum/src/domain/record/shadowRootsController.ts
@@ -1,8 +1,9 @@
 import { addTelemetryDebug, DOM_EVENT, isExperimentalFeatureEnabled } from '@datadog/browser-core'
 import type { RumConfiguration } from '@datadog/browser-rum-core'
+import type { InputCallback } from './observers/inputObserver'
+import { initInputObserver } from './observers/inputObserver'
 import { startMutationObserver } from './observers/mutationObserver'
-import { initInputObserver } from './observers/observers'
-import type { MutationCallBack, InputCallback } from './observers/observers'
+import type { MutationCallBack } from './observers/observers'
 
 interface ShadowRootController {
   stop: () => void
